# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FarmaSystem - API de Gestão de Farmácia
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: Listar clientes
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Criar cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Obter cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Atualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Excluir cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Excluído
  /customers/{id}/sales:
    get:
      operationId: getCustomerSales
      tags: [customers]
      summary: Histórico de compras do cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Sale"

  # ─── Suppliers ────────────────────────────────────────────────
...
  /prescriptions:
    get:
      operationId: listPrescriptions
      tags: [prescriptions]
      summary: Listar receitas
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prescription"
    post:
      operationId: createPrescription
      tags: [prescriptions]
      summary: Registrar receita
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrescriptionBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"
  /prescriptions/{id}:
    get:
      operationId: getPrescription
      tags: [prescriptions]
      summary: Obter receita
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prescription"

  # ─── Controlled Medications ───────────────────────────────────
  /controlled:
    get:
      operationId: listControlledDispensations
      tags: [controlled]
      summary: Listar dispensações de controlados
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ControlledDispensation"
    post:
      operationId: createControlledDispensation
      tags: [controlled]
      summary: Registrar dispensação de controlado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateControlledDispensationBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlledDispensation"
  /controlled/{id}:
    get:
      operationId: getControlledDispensation
      tags: [controlled]
      summary: Obter dispensação de controlado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ControlledDispensation"

  # ─── Purchases ────────────────────────────────────────────────
...
  /financial/transactions:
    get:
      operationId: listFinancialTransactions
      tags: [financial]
      summary: Listar transações financeiras
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinancialTransaction"
    post:
      operationId: createFinancialTransaction
      tags: [financial]
      summary: Criar transação financeira
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancialTransactionBody"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancialTransaction"
...
components:
  schemas:
...
    Sale:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerCpf:
          type: ["string", "null"]
        userId:
          type: integer
        userName:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/SaleItem"
        subtotal:
          type: number
        discount:
          type: number
        total:
          type: number
        paymentMethod:
          type: string
          enum: [dinheiro, cartao_credito, cartao_debito, pix, convenio]
        status:
          type: string
          enum: [concluida, cancelada, pendente]
        createdAt:
          type: string
      required:
        - id
        - userId
        - userName
        - items
        - subtotal
        - discount
        - total
        - paymentMethod
        - status
        - createdAt
    SaleItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        total:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - unitPrice
        - discount
        - total
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        cpf:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        loyaltyPoints:
          type: integer
        totalPurchases:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - loyaltyPoints
        - totalPurchases
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        cpf:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        birthDate:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required:
        - name
...
    Prescription:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        doctorName:
          type: string
        doctorCrm:
          type: string
        prescriptionDate:
          type: string
        expirationDate:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrescriptionItem"
        status:
          type: string
          enum: [pendente, dispensada, vencida, cancelada]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - doctorName
        - doctorCrm
        - prescriptionDate
        - items
        - status
        - createdAt
    PrescriptionItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        dosageInstructions:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - quantity
    CreatePrescriptionBody:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        doctorName:
          type: string
        doctorCrm:
          type: string
        prescriptionDate:
          type: string
        expirationDate:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PrescriptionItem"
        notes:
          type: ["string", "null"]
      required:
        - doctorName
        - doctorCrm
        - prescriptionDate
        - items
    ControlledDispensation:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerCpf:
          type: string
        productId:
          type: integer
        productName:
          type: string
        prescriptionId:
          type: ["integer", "null"]
        quantity:
          type: integer
        anvisaClass:
          type: string
        notificationNumber:
          type: ["string", "null"]
        retentionRequired:
          type: boolean
        retained:
          type: boolean
        dispensedBy:
          type: string
        createdAt:
          type: string
      required:
        - id
        - customerId
        - customerName
        - customerCpf
        - productId
        - productName
        - quantity
        - anvisaClass
        - retentionRequired
        - retained
        - dispensedBy
        - createdAt
    CreateControlledDispensationBody:
      type: object
      properties:
        customerId:
          type: integer
        productId:
          type: integer
        prescriptionId:
          type: ["integer", "null"]
        quantity:
          type: integer
        anvisaClass:
          type: string
        notificationNumber:
          type: ["string", "null"]
        retentionRequired:
          type: boolean
        retained:
          type: boolean
      required:
        - customerId
        - productId
        - quantity
        - anvisaClass
        - retentionRequired
        - retained
...
    FinancialTransaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [receita, despesa]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
        paidDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [pendente, pago, vencido, cancelado]
        createdAt:
          type: string
      required:
        - id
        - type
        - category
        - description
        - amount
        - dueDate
        - status
        - createdAt
    CreateFinancialTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [receita, despesa]
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        dueDate:
          type: string
      required:
        - type
        - category
        - description
        - amount
        - dueDate
...